restructured sourcecode for extendibility
