+++ RobotLog2RQM/logging.py
@@ -1,0 +1,179 @@
+#  Copyright 2020-2023 Robert Bosch GmbH
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+# ******************************************************************************
+#
+# File: robotlog2rqm.py
+#
+# Initially created by Tran Duy Ngoan(RBVH/EMC51) / October 2025
+#
+# Logging class
+#
+# History:
+#
+# 2025-10-20:
+#  - initial version
+#
+# ******************************************************************************
+import sys
+import colorama as col
+
+class Logger():
+   """
+Logger class for logging message.
+   """
+   output_logfile = None
+   output_console = True
+   color_normal   = col.Fore.WHITE + col.Style.NORMAL
+   color_error    = col.Fore.RED + col.Style.BRIGHT
+   color_warn     = col.Fore.YELLOW + col.Style.BRIGHT
+   color_reset    = col.Style.RESET_ALL + col.Fore.RESET + col.Back.RESET
+   prefix_warn    = "WARN: "
+   prefix_error   = "ERROR: "
+   prefix_fatalerror = "FATAL ERROR: "
+   prefix_all = ""
+   dryrun = False
+
+   @classmethod
+   def config(cls, output_console=True, output_logfile=None, dryrun=False):
+      """
+Configure Logger class.
+
+**Arguments:**
+
+*  ``output_console``
+
+   / *Condition*: optional / *Type*: bool / *Default*: True /
+
+   Write message to console output.
+
+*  ``output_logfile``
+
+   / *Condition*: optional / *Type*: str / *Default*: None /
+
+   Path to log file output.
+
+*  ``dryrun``
+
+   / *Condition*: optional / *Type*: bool / *Default*: True /
+
+   If set, a prefix as 'dryrun' is added for all messages.
+
+**Returns:**
+
+(*no returns*)
+      """
+      cls.output_console = output_console
+      cls.output_logfile = output_logfile
+      cls.dryrun = dryrun
+      if cls.dryrun:
+         cls.prefix_all = cls.color_warn + "DRYRUN  " + cls.color_reset
+
+   @classmethod
+   def log(cls, msg='', color=None, indent=0):
+      """
+Write log message to console/file output.
+
+**Arguments:**
+
+*  ``msg``
+
+   / *Condition*: optional / *Type*: str / *Default*: '' /
+
+   Message which is written to output.
+
+*  ``color``
+
+   / *Condition*: optional / *Type*: str / *Default*: None /
+
+   Color style for the message.
+
+*  ``indent``
+
+   / *Condition*: optional / *Type*: int / *Default*: 0 /
+
+   Offset indent.
+
+**Returns:**
+
+(*no returns*)
+      """
+      if color==None:
+         color = cls.color_normal
+      if cls.output_console:
+         print(cls.prefix_all + cls.color_reset + color + " "*indent + msg + cls.color_reset)
+      if cls.output_logfile!=None and os.path.isfile(cls.output_logfile):
+         with open(cls.output_logfile, 'a') as f:
+            f.write(" "*indent + msg)
+      return
+
+   @classmethod
+   def log_warning(cls, msg, indent=0):
+      """
+Write warning message to console/file output.
+
+**Arguments:**
+
+*  ``msg``
+
+   / *Condition*: required / *Type*: str /
+
+   Warning message which is written to output.
+
+*  ``indent``
+
+   / *Condition*: optional / *Type*: int / *Default*: 0 /
+
+   Offset indent.
+
+**Returns:**
+
+(*no returns*)
+      """
+      cls.log(cls.prefix_warn+str(msg), cls.color_warn, indent)
+
+   @classmethod
+   def log_error(cls, msg, fatal_error=False, indent=0):
+      """
+Write error message to console/file output.
+
+*  ``msg``
+
+   / *Condition*: required / *Type*: str /
+
+   Error message which is written to output.
+
+*  ``fatal_error``
+
+   / *Condition*: optional / *Type*: bool / *Default*: False /
+
+   If set, tool will terminate after logging error message.
+
+*  ``indent``
+
+   / *Condition*: optional / *Type*: int / *Default*: 0 /
+
+   Offset indent.
+
+**Returns:**
+
+(*no returns*)
+      """
+      prefix = cls.prefix_error
+      if fatal_error:
+         prefix = cls.prefix_fatalerror
+
+      cls.log(prefix+str(msg), cls.color_error, indent)
+      if fatal_error:
+         cls.log(f"{sys.argv[0]} has been stopped!", cls.color_error)
+         exit(1)
